requirements.txt
Flask==3.0.0
boto3==1.34.0
botocore==1.34.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
